Maximum Depth of Binary Tree
Maximum Depth of Binary Tree
Given a binary tree, the task is to find the maximum depth of the tree. The maximum depth or height of the tree is the number of edges in the tree from the root to the deepest node.
Using Recursion - O(n) Time and O(h) Space
The idea is to recursively calculate the height of the left and the right subtrees of a node and then find height to the node as max of the heights of two children plus 1.

Algorithm:

If the tree is empty, return -1.
Otherwise, do the following:
Get the height of the left subtree recursively, i.e., call height(node->left).
Get the height of the right subtree recursively, i.e., call height(node->right).
Compute the maximum of the heights of the left and right subtrees and add 1 to it for the current node.
height = max(height of left subtree, height of right subtree) + 1.
Return the height.