k-th Smallest in BST
k-th Smallest in BST (Order Statistics in BST)
Given a Binary Search Tree (BST) and a positive integer k, find the kth smallest element in the Binary Search Tree. 

Example:

Input: k = 3
Output: 10
Explanation: The inorder traversal of given BST is [4, 8, 10, 12, 14, 20, 22] and its 3rd smallest element is 10.

[Expected Approach] Using In-Order traversal - O(k) Time and O(h) Space
The idea is to traverse the binary search tree using in-order traversal while maintaining the count of nodes traversed. If the node count becomes equal to k, then return the node.
We can optimize the space using Morris Traversal. Refer K’th smallest element in BST using O(1) Extra Space for details.

[Alternate Approach] Using Augmented Tree - O(h) Time and O(h) Space
The idea is to include a new data member ''lCount for each node. We can keep track of elements in the left subtree of every node while building the tree and storing it in this new data member.
Assuming that the root is having 'lCount' nodes in its left subtree. If k is equal to lCount + 1, then root is k-th node. If k is less than lCount + 1, we will continue our search (recursion) for the kth smallest element in the left subtree of root. If k is greater than lCount + 1, we continue our search in the right subtree. Note that we need the count of elements in the left subtree only.