k largest elements in an array
Find k largest elements in an array
Given an array arr[] and an integer k, the task is to find k largest elements in the given array. Elements in the output array should be in decreasing order.

Examples:

Input:  [1, 23, 12, 9, 30, 2, 50], k = 3
Output: [50, 30, 23]

Input:  [11, 5, 12, 9, 44, 17, 2], k = 2
Output: [44, 17]

Table of Content

[Naive Approach] Using Sorting
[Expected Approach] Using Priority Queue(Min-Heap)
[Alternate Approach] Using Quick Select algorithm - O(n^2) in Worst
[Naive Approach] Using Sorting
The idea is to sort the input array in descending order, so the first k elements in the array will be the k largest elements.
// Java program to find k largest elements in an array using
// sorting

import java.util.*;

class GfG {
    static ArrayList<Integer> kLargest(int[] arr, int k) {
      	int n = arr.length;
      
         // Convert int type to Integer 
      	// for sorting with a comparator
        Integer[] arrInteger = 
          	Arrays.stream(arr).boxed().toArray(Integer[]::new);

        // Sort the array in descending order
        Arrays.sort(arrInteger, Collections.reverseOrder());

        // Store the first k elements in result list
        ArrayList<Integer> res = new ArrayList<>();
        for (int i = 0; i < k; i++) 
            res.add(arrInteger[i]);
          
        return res;
    }

    public static void main(String[] args) {
        int[] arr = {1, 23, 12, 9, 30, 2, 50};
        int k = 3;

        ArrayList<Integer> res = kLargest(arr, k);
        for (int ele : res)
            System.out.print(ele + " ");
    }
}


Output
50 30 23 
Time complexity: O(n * log n)
Auxiliary Space: O(1)

[Expected Approach] Using Priority Queue(Min-Heap)
The idea is, as we iterate through the array, we keep track of the k largest elements at each step. To do this, we use a min-heap. First, we insert the initial k elements into the min-heap. After that, for each next element, we compare it with the top of the heap. Since the top element of the min-heap is the smallest among the k elements, if the current element is larger than the top, it means the top element is no longer one of the k largest elements. In this case, we remove the top and insert the larger element. After completing the entire traversal, the heap will contain exactly the k largest elements of the array.

Output
50 30 23 
Time Complexity: O(n * log k), this solution can work in O(k + (n-k) Log K) as build heap take linear time.
Auxiliary Space: O(k)

Note: JavaScript does not seem to support min heap in native implementation, so it is recommended to use Quick Select implementation.

[Alternate Approach] Using Quick Select algorithm
The idea is to use the partitioning step of QuickSort to find the k largest elements in the array, without sorting the entire array. When sorting the element in the descending order, the partitioning step rearranges the elements in a way that all elements greater than or equal to a chosen pivot (usually the last element) are placed on its left, and all elements lesser than the pivot are on its right. And pivot element in its correct sorted position.
After each partition, we compare the number of elements in the leftpart of the array (which contains all elements greater than or equal to the pivot) with k:

Number of elements in the left = k, it means all elements in the left part (including pivot) are the k largest elements.
Number of elements in the left > k, it means that k largest elements exist in the left subarray only, so we recursively search in the left subarray.
Number of elements in the left < k, it means that the k largest elements include the entire left part of the array along with some elements from the right part. Therefore we reduce k by the number of elements already covered on the left side and search in the right subarray.

// Java program to find the k largest elements in the array
// using partitioning step of quick sort

import java.util.*;

class GfG {

    // Function to partition the array around a pivot
    static int partition(int[] arr, int left, int right) {
        
        // Last element is chosen as a pivot.
        int pivot = arr[right];
        int i = left;

        for (int j = left; j < right; j++) {
          
          	// Elements greater than or equal to pivot
      		// are placed in the left side of pivot
            if (arr[j] >= pivot) {
                int temp = arr[i];
                arr[i] = arr[j];
                arr[j] = temp;
                i++;
            }
        }
        
        int temp = arr[i];
        arr[i] = arr[right];
        arr[right] = temp;
        
        // The correct sorted position of the pivot
        return i;
    }

    static void quickSelect(int[] arr, int left, int right, int k) {
        if (left <= right) {
            int pivotIdx = partition(arr, left, right);
            
            // Count of all elements in the left part
            int leftCnt = pivotIdx - left + 1;
          
            // If leftCnt is equal to k, then we have
      		// found the k largest element
            if (leftCnt == k) 
                return;
            
            // Search in the left subarray
            if (leftCnt > k)
                quickSelect(arr, left, pivotIdx - 1, k);
                
            // Reduce the k by number of elements already covered
            // and search in the right subarray
            else 
                quickSelect(arr, pivotIdx + 1, right, k - leftCnt);
        }
    }

    static ArrayList<Integer> kLargest(int[] arr, int k) {
        quickSelect(arr, 0, arr.length - 1, k);
        ArrayList<Integer> res = new ArrayList<>();
      
        // First k elements of the array, will be the largest
      	for(int i = 0; i < k; i++)
          	res.add(arr[i]);
      
      	// Sort the result in descending order
      	Collections.sort(res, Collections.reverseOrder());
        return res;
    }

    public static void main(String[] args) {
        int[] arr = {1, 23, 12, 9, 30, 2, 50};
        int k = 3;

        ArrayList<Integer> res = kLargest(arr, k);
        for (int ele : res)
            System.out.print(ele + " ");
    }
}